# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Servana by Citrus API",
        "version": "v1",
        "description": "Production API contract generated by dedoc/scramble (authoritative) via the Servana wrapper. Do not edit by hand — run `composer api:openapi`."
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/v1/auth/magic-link": {
            "post": {
                "operationId": "auth.magic-link.request",
                "summary": "POST /auth/magic-link — always 202, never reveals account existence",
                "tags": [
                    "MagicLink"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestMagicLinkRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "If the email exists and is active, a link was sent."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "429": {
                        "description": "Rate limited",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
        "/api/v1/files/{uploadedFile}/download": {
            "get": {
                "operationId": "files.download",
                "tags": [
                    "File"
                ],
                "parameters": [
                    {
                        "name": "uploadedFile",
                        "in": "path",
                        "required": true,
                        "description": "The uploaded file ulid",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "403": {
                        "description": "Permission denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "429": {
                        "description": "Rate limited",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        },
        "schemas": {
...
            "ErrorEnvelope": {
                "type": "object",
                "description": "Standard API error envelope (Plan §11.5).",
                "properties": {
                    "error": {
                        "type": "object",
                        "properties": {
                            "code": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            },
                            "fields": {
                                "type": "object"
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "required": [
                            "code",
                            "message"
                        ]
                    }
                },
                "required": [
                    "error"
                ]
            },
...
            "RequestMagicLinkRequest": {
                "type": "object",
                "description": "Validates a Magic Link request (Plan §9.1). Public endpoint — no auth.",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    }
                },
                "required": [
                    "email"
                ],
                "title": "RequestMagicLinkRequest"
            },
...
        },
        "securitySchemes": {
            "sanctumSession": {
                "type": "apiKey",
                "in": "cookie",
                "name": "servana_session",
                "description": "Laravel Sanctum first-party stateful session cookie + X-XSRF-TOKEN."
            }
        }
    }
}